Remove final reshapes even for Quartus
